+++ pkg/parsergen/backend/cpp/doc.go
@@ -1,0 +1,2 @@
+// Package cpp provides a parser generator backend which takes a parse table as input and outputs C++ source code.
+package cpp