+++ pkg/scannergen/backend/cpp/doc.go
@@ -1,0 +1,2 @@
+// Package cpp provides a scanner generator backend which takes a DFA as input and outputs C++ source code.
+package cpp